Update module to 2016-11-02 Gateway release
