+++ modules/modbus_read/inc/modbus_read.h
@@ -12,7 +12,7 @@ extern "C"
 {
 #endif
 
-MODULE_EXPORT void MODULE_STATIC_GETAPIS(MODBUSREAD_MODULE)(MODULE_APIS* apis);
+MODULE_EXPORT const MODULE_API* MODULE_STATIC_GETAPI(MODBUSREAD_MODULE)(const MODULE_API_VERSION gateway_api_version);
 
 #ifdef __cplusplus
 }
