--- modules/modbus_read/inc/modbus_read_hl.h
@@ -1,21 +1,0 @@
-// Copyright (c) Microsoft. All rights reserved.
-// Licensed under the MIT license. See LICENSE file in the project root for full license information.
-
-#ifndef MODBUS_READ_HL_H
-#define MODBUS_READ_HL_H
-
-#include "module.h"
-
-#ifdef __cplusplus
-extern "C"
-{
-#endif
-
-MODULE_EXPORT void MODULE_STATIC_GETAPIS(MODBUSREAD_MODULE_HL)(MODULE_APIS* apis);
-
-#ifdef __cplusplus
-}
-#endif
-
-
-#endif /*MODBUS_READ_HL_H*/
