+++ modules/modbus_read/tests/CMakeLists.txt
@@ -2,8 +2,5 @@
 #Licensed under the MIT license. See LICENSE file in the project root for full license information.
 
 cmake_minimum_required(VERSION 2.8.11)
-#this is CMakeLists for modbus read tests
 
 add_subdirectory(modbus_read_unittests)
-add_subdirectory(modbus_read_hl_unittests)
-
