--- modules/modbus_read/tests/modbus_read_hl_unittests/main.c
@@ -1,11 +1,0 @@
-// Copyright (c) Microsoft. All rights reserved.
-// Licensed under the MIT license. See LICENSE file in the project root for full license information.
-
-#include "testrunnerswitcher.h"
-
-int main(void)
-{
-    size_t failedTestCount = 0;
-    RUN_TEST_SUITE(modbus_read_hl_unittests, failedTestCount);
-    return failedTestCount;
-}
