+++ samples/modbus/src/main.c
@@ -16,7 +16,7 @@ int main(int argc, char** argv)
     }
     else
     {
-        if ((gateway = Gateway_Create_From_JSON(argv[1])) == NULL)
+        if ((gateway = Gateway_CreateFromJson(argv[1])) == NULL)
         {
             printf("failed to create the gateway from JSON\n");
         }
@@ -25,7 +25,7 @@ int main(int argc, char** argv)
             printf("gateway successfully created from JSON\n");
             printf("gateway shall run until ENTER is pressed\n");
             (void)getchar();
-            Gateway_LL_Destroy(gateway);
+            Gateway_Destroy(gateway);
         }
     }
 	return 0;
